fix: restructured search math and strings subpackage
