+++ strings/ahocorasick/advancedahocorasick_test.go
@@ -1,9 +1,7 @@
-package advancedahocorasick
+package ahocorasick
 
 import (
-	"fmt"
 	"reflect"
-	"strings"
 	"testing"
 )
 
@@ -50,7 +48,7 @@ var testCases = []struct {
 	},
 }
 
-func TestAhoCorasick(t *testing.T) {
+func TestAdvanced(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			actual := AhoCorasick(tc.text, tc.words)
@@ -62,22 +60,4 @@ func TestAhoCorasick(t *testing.T) {
 			}
 		})
 	}
-}
-
-func convertToString(res Result) string {
-	// var r string
-	var r strings.Builder
-	for key, val := range res.occurrences {
-		r.WriteString(fmt.Sprintf("Word: '%s' at positions: ", key))
-		for i := range val {
-			r.WriteString(fmt.Sprintf("%d", val[i]))
-			if i != len(val)-1 {
-				// r = r + fmt.Sprintf(", ")
-				r.WriteString(", ")
-			}
-		}
-		// r = r + fmt.Sprintf(". ")
-		r.WriteString(". ")
-	}
-	return r.String()
 }
