+++ strings/ahocorasick/patterns.txt
