+++ strings/ahocorasick/text.txt
