+++ strings/bom/bom.go
