+++ strings/horspool/horspool.go
