+++ strings/kmp/kmp.go
