+++ strings/kmp/kmp_test.go
