+++ strings/search/patternsearch.go
@@ -8,9 +8,9 @@ Time Complexity : O(n*m)
     m = length of pattern
 */
 
-package naive_string_search
+package search
 
-func naivePatternSearch(text string, pattern string) []int {
+func Naive(text string, pattern string) []int {
 	var positions []int
 	for i := 0; i <= len(text)-len(pattern); i++ {
 		var match bool = true
