--- strings/single_string_matching/pattern.txt
@@ -1,1 +1,0 @@
-announce
