--- strings/single_string_matching/text.txt
@@ -1,1 +1,0 @@
-CPM_annual_conference_announce
